requirements.txt
numpy>=1.26
pandas>=2.2
PyYAML>=6.0
openpyxl>=3.1
opencv-python>=4.9
matplotlib>=3.8
Pillow>=10.2
streamlit>=1.33
tensorflow>=2.15
